requirements.txt
ExifRead==3.0.0
Flask==2.0.1
folium==0.14.0
gpsphoto==2.2.3
loguru==0.7.0
piexif==1.1.3
Pillow==9.5.0
requests==2.25.1
